Fix compilation problem with CarpetHDF5 reader.
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@17300 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/databases/CarpetHDF5/avtCarpetHDF5FileFormat.h
@@ -111,8 +111,8 @@ class dataset_entry
                     
                     if (_ghosts[2*d] > 0 || _ghosts[2*d+1] > 0)
                     {
-                       if (_rl > 0 && _iorigin_ghosted[d] % int(pow(2, 1)) >  0) { _npoints_ghosted[d] += 1; _iorigin_ghosted[d] -= 1; _origin_ghosted[d] -= _delta[d]; _ghosts[2*d] -= 1; }
-                       if (_rl > 0 && _npoints_ghosted[d] % int(pow(2, 1)) == 0) { _npoints_ghosted[d] -= 1; /*_ghosts[2*d+1 += 1;*/ }
+                       if (_rl > 0 && _iorigin_ghosted[d] % int(pow((double)2, 1)) >  0) { _npoints_ghosted[d] += 1; _iorigin_ghosted[d] -= 1; _origin_ghosted[d] -= _delta[d]; _ghosts[2*d] -= 1; }
+                       if (_rl > 0 && _npoints_ghosted[d] % int(pow((double)2, 1)) == 0) { _npoints_ghosted[d] -= 1; /*_ghosts[2*d+1 += 1;*/ }
                     }
                 }
                 
@@ -158,8 +158,8 @@ class dataset_entry
                     
                     if (_ghosts[2*d] > 0 || _ghosts[2*d+1] > 0)
                     {
-                       if (_rl > 0 && _iorigin_ghosted[d] % int(pow(2, 1)) >  0) { _npoints_ghosted[d] += 1; _iorigin_ghosted[d] -= 1; _origin_ghosted[d] -= _delta[d]; _ghosts[2*d] -= 1; }
-                       if (_rl > 0 && _npoints_ghosted[d] % int(pow(2, 1)) == 0) { _npoints_ghosted[d] -= 1; /*_ghosts[2*d+1 += 1;*/ }
+                       if (_rl > 0 && _iorigin_ghosted[d] % int(pow((double)2, 1)) >  0) { _npoints_ghosted[d] += 1; _iorigin_ghosted[d] -= 1; _origin_ghosted[d] -= _delta[d]; _ghosts[2*d] -= 1; }
+                       if (_rl > 0 && _npoints_ghosted[d] % int(pow((double)2, 1)) == 0) { _npoints_ghosted[d] -= 1; /*_ghosts[2*d+1 += 1;*/ }
                     }
                 }
